1. initial support for rich text in xssf" 2. cleaned common ooxml-ole2 interfaces, removed ole2-specific stuff 3. added new examples from the quick guide 4. misc xssf refactoring, tending to use enums intstead of final static constants 5. refactored XSSFFont 6. included test-ooxml and jar-ooxml in the dist target, they are part of release and should be there
git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@696584 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/interfaces-jdk14/org/apache/poi/ss/usermodel/Comment.java
@@ -18,6 +18,6 @@
 package org.apache.poi.ss.usermodel;
 
 public interface Comment {
-	public void setRow(short row);
+	public void setRow(int row);
 	public void setColumn(short row);
 }
